initial commit
This is a Javascript Form Validation free project. In this project, we will check whether the username, email and password entered are correct or incorrect. We will show an error if the details entered are incorrect. And if the details entered are correct then we will show the success message.